backend/requirements.txt
# Modern FastAPI + Pydantic v2 (Python 3.13 compatible)
fastapi==0.115.0
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.35  # Python 3.13 compatible
pymysql==1.1.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic>=2.10.0  # Python 3.13 precompiled wheels
pydantic-settings>=2.7.0  # Python 3.13 compatible
reportlab==4.0.7
python-dateutil==2.8.2
gunicorn==21.2.0

# Security & Rate Limiting
slowapi==0.1.9
python-dotenv==1.0.0

# Database Migrations
alembic==1.17.1

# Testing & Coverage
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
httpx>=0.27.0  # Required by TestClient
faker>=28.0.0  # For generating test data

# Monitoring & Observability
psutil>=6.1.0  # System metrics
sentry-sdk[fastapi]>=2.21.0  # Optional error tracking
